regression tests command & rename things & slight makefile cleanup
