+++ app/commands/generate_spec/command.py
