+++ app/commands/regression_tests/__init__.py
@@ -1,0 +1,3 @@
+from app.commands.regression_tests.command import RegressionTestsCommand
+
+__all__ = ["RegressionTestsCommand"]
