+++ app/commands/runserver/__init__.py
@@ -1,3 +1,3 @@
-from app.commands.runserver.start import RunServerCommand
+from app.commands.runserver.command import RunServerCommand
 
 __all__ = ["RunServerCommand"]
