+++ main.py
@@ -3,9 +3,9 @@ import os
 import click
 
 import app.commands.generate_spec as generate_spec_cmd
+import app.commands.regression_tests as regression_tests_command
 import app.commands.runserver as runserver_cmd
 from app.lib import commands
-from tests.regression import upload_simple_table
 
 
 @click.group()
@@ -22,22 +22,18 @@ def cli():
     help="Path to configuration file",
 )
 def runserver(config: str):
-    commands.run(
-        runserver_cmd.RunServerCommand(config),
-    )
+    commands.run(runserver_cmd.RunServerCommand(config))
 
 
 @cli.command(short_help="Generate OpenAPI spec and write it to file")
 @click.option("-o", "--output", type=str, required=True, help="Where to put resulting JSON")
 def generate_spec(output: str):
-    commands.run(
-        generate_spec_cmd.GenerateSpecCommand(output),
-    )
+    commands.run(generate_spec_cmd.GenerateSpecCommand(output))
 
 
 @cli.command(short_help="Run regression tests")
 def regression_tests():
-    upload_simple_table.run()
+    commands.run(regression_tests_command.RegressionTestsCommand())
 
 
 if __name__ == "__main__":
